Documentation: contributing: Update Source references
The libcamera project is no longer in 'early' stages of development, and
we do make releases.

Update the Source Code section to reflect this reality and report on the
current expectations of release numbering schemes.

Furthermore update the development mirror as hosted on Freedesktop and
deprecate the LinuxTV mirror, as Freedesktop also hosts our release tar
balls and CI infrastructure.

Reviewed-by: Umang Jain <uajain@igalia.com>
Reviewed-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>